Refactor model from core to feature
