+++ app/src/main/java/com/example/tasky/agenda/data/EventRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.example.tasky.agenda.data
 import com.example.tasky.agenda.data.local.AgendaDao
 import com.example.tasky.agenda.data.mapper.toDomain
 import com.example.tasky.agenda.domain.EventRepository
-import com.example.tasky.core.domain.model.AgendaItem
+import com.example.tasky.agenda.domain.model.AgendaItem
 import javax.inject.Inject
 
 class EventRepositoryImpl @Inject constructor(
