+++ app/src/main/java/com/example/tasky/agenda/data/mapper/AttendeeMapper.kt
@@ -2,7 +2,7 @@ package com.example.tasky.agenda.data.mapper
 
 import com.example.tasky.agenda.data.local.model.AttendeeEntity
 import com.example.tasky.agenda.data.remote.model.AttendeeResponse
-import com.example.tasky.core.domain.model.Attendee
+import com.example.tasky.agenda.domain.model.Attendee
 
 fun AttendeeResponse.toDomain() =
     Attendee(
