+++ app/src/main/java/com/example/tasky/agenda/domain/EventRepository.kt
@@ -1,6 +1,6 @@
 package com.example.tasky.agenda.domain
 
-import com.example.tasky.core.domain.model.AgendaItem
+import com.example.tasky.agenda.domain.model.AgendaItem
 
 interface EventRepository {
     suspend fun getEventById(eventId: String): AgendaItem.Event
