+++ app/src/main/java/com/example/tasky/agenda/presentation/home/AgendaScreen.kt
@@ -33,6 +33,7 @@ import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.tasky.R
+import com.example.tasky.agenda.domain.model.AgendaItem
 import com.example.tasky.agenda.domain.model.Day
 import com.example.tasky.agenda.presentation.home.components.DayPill
 import com.example.tasky.agenda.presentation.home.components.EventItem
@@ -44,7 +45,6 @@ import com.example.tasky.agenda.presentation.home.components.TaskItem
 import com.example.tasky.agenda.presentation.home.components.TaskyDropdownMenu
 import com.example.tasky.agenda.presentation.home.viewmodel.AgendaEvent
 import com.example.tasky.agenda.presentation.home.viewmodel.AgendaState
-import com.example.tasky.core.domain.model.AgendaItem
 import com.example.tasky.ui.theme.Black
 import com.example.tasky.ui.theme.TaskyTheme
 import com.vanpra.composematerialdialogs.MaterialDialog
