+++ app/src/main/java/com/example/tasky/agenda/presentation/itemdetail/components/DetailAttendeeList.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.tasky.core.domain.model.Attendee
+import com.example.tasky.agenda.domain.model.Attendee
 
 @Composable
 fun DetailAttendeeList(
