+++ app/src/main/java/com/example/tasky/agenda/presentation/itemdetail/components/VisitorRow.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.tasky.R
-import com.example.tasky.core.domain.model.Attendee
+import com.example.tasky.agenda.domain.model.Attendee
 import com.example.tasky.ui.theme.DarkGray
 import com.example.tasky.ui.theme.Light2
 import com.example.tasky.ui.theme.LightBlue2
